Bugfix for Issue #2
Cause of bug: al_init_primitives_addon() was not called. On Linux, this happened not to cause a problem, but due to time constraints the Windows build wasn't thoroughly tested, and if I recall correctly the primitives add-on is only used on the build screen when there's not enough money to build the most expensive building.